requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
requests==2.31.0
pytest==7.4.4
behave==1.2.6
python-dotenv
